This adds saving items to disk using wkhtmltopdf
This is a quick test to see how well saving to PDF works. The good news
is: pretty well actually.
Saving an item stores it in a file specified as:
<cwd>/saved/SITENAME/ITEMTITLE.pdf
Doing so also sends a message back to the user with a link that they can
click on to view or download to mail on to friends.
To do:
1) see if the PDF's are still text searchable
2) push a change made to the PDF wrapper for wkhtmltopdf that works
around wkhtmltopdf sometimes lying about how successful it was in
generating the PDF.