Fixes all-read bug
Even though the main content had been removed, waypoints still existed.
So when a single item was shown, it would replace the main content
leaving the waypoints, and when another call to loadItems() was made,
all the waypoints would be triggered in a weird loop.

Now to make the "next page" waypoint stop triggering early...