Fix WkHtmlToPdf save bug
Fixed the non-existing logic in the change I made to WkHtmlToPdf.php.
Now it tries to save the page, if it exits with zero everything is cool.
If it's a non-zero return code, if the code is 2 AND a file exists,
chances are that WkHtmlToPdf made a file just fine but reported an error
about a background image or something, so return true.
If code is 2 and no file exists, then we're still in an error condition.